[woogie.pyun] 인증관리 - Step1 & 2 (#7)
* docs: add README.md

* docs: add application.yml

* feat: add AuthController

* docs: add application.yml

* fix: fix sql
- create only if table is not exists
- truncate for test

* feat: add method that matches the username, password

* feat: add method that validate tokenRequest

* test: truncate table for tests

* docs: check the list
- create token api

* fix: add dependency
- glassfish.jaxb

* fix: add empty contructor

* feat: add method that find member and response

* feat: add method that findUserByName

* feat: add request extractor

* feat: check step 1

* feat: add ArgumentResolver related codes

* refactor: add truncate logic to all tests

* refactor: make Resolver throw exception when header is null

* refactor: add login requirements to ReservationController

* fix: fix trivial issues in tests and Dtos

* refactor: add token validation to argumentResolver

* refactor: make JWT hold memberId not username and change affected

* docs: edit README.md

* refactor: remove unused method

* refactor: remove unnecessary annotation field

* refactor: make ArgumentResolver return Member object

* refactor: split methods and move token generating responsibility to service

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,8 +1,11 @@
 package nextstep.reservation;
 
+import nextstep.member.Member;
+import nextstep.member.MemberDao;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
 import nextstep.support.DuplicateEntityException;
+import nextstep.support.NotExistEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
@@ -14,14 +17,16 @@ public class ReservationService {
     public final ReservationDao reservationDao;
     public final ThemeDao themeDao;
     public final ScheduleDao scheduleDao;
+    public final MemberDao memberDao;
 
-    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao) {
+    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao, MemberDao memberDao) {
         this.reservationDao = reservationDao;
         this.themeDao = themeDao;
         this.scheduleDao = scheduleDao;
+        this.memberDao = memberDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(Member loggedInMember, ReservationRequest reservationRequest) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
             throw new NullPointerException();
@@ -34,7 +39,7 @@ public class ReservationService {
 
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                loggedInMember.getName()
         );
 
         return reservationDao.save(newReservation);
@@ -49,10 +54,14 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Member loggedInMember, Long id) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
+        }
+
+        if (!reservation.getName().equals(loggedInMember.getName())) {
+            throw new NotExistEntityException();
         }
 
         reservationDao.deleteById(id);
